D-LK10-ENTRY-TRAMP-PROLOGUE audit fold: 8-agent review on 6a2eeec
Standing 8-agent review (silent-failure + code-reviewer + comment-analyzer
+ code-architect + code-simplifier + type-design + pr-test-analyzer +
Explore cross-plan) on the runnable-CLI-binary substrate (entryStackPointerBias
field + alignedSizeWithBias helper + trampoline prologue emission).

**Strong 2-agent convergence** (type-design C1 + silent-failure HIGH-1):
the `hasAbiInfo` predicate at target_schema.cpp:929-934 omitted
`entryStackPointerBias != 0`. A cc with ONLY entryStackPointerBias=7 and
all other ABI fields zeroed silently passed validation (the `< stackAlignment`
check sits inside the hasAbiInfo arm). Folded: predicate now includes
the new field so the validator enters its arm and rejects via the
stackAlignment-pow2 check.

**FOLD-NOW items** (all 7 ≥80-confidence findings folded in-slice):

1. code-simplifier FN-1: dropped 3 redundant `static_cast<uint32_t>`
   wrappers at entry_trampoline.cpp's alignedSizeWithBias call site —
   uint16→uint32 is integral promotion, the casts read like narrowing.

2. comment-analyzer HIGH: `sysv_x64` typo in entryStackPointerBias
   docblock — canonical cc name is `sysv_amd64` (verified via grep:
   zero `sysv_x64` occurrences anywhere in src/). Fixed.

3. code-reviewer Important-1 + comment-analyzer minor: phantom
   `processExit.preCallStackAdjustBytes` reference in
   entry_trampoline.cpp:147-150 docblock (left over from the reverted
   first-attempt approach). Rewritten to cite the actual cc-field
   inputs to alignedSizeWithBias.

4. code-architect FOLD-NOW: D-LK10-ENTRY-SYNTH-PREFIX-SCHEMA anchor
   citation was missing from production source at entry_trampoline.cpp:53
   (the hardcoded "sym_" / "_sym_" prefixes). Added inline anchor +
   closure trigger comment per 38ac5a2 discipline.

5. type-design C1 + silent-failure HIGH-1 (2-agent): hasAbiInfo
   predicate fix described above.

6. **pr-test-analyzer Gaps 1-6 + silent-failure HIGH-3 + code-reviewer
   Important-2 — all addressed via 15 new tests**:
   - test_lir_callconv.cpp: 10 alignedSizeWithBias unit tests covering
     the formula matrix (Win64 32+8→40; SysV/ARM64 0+0→0; bias-only;
     fast-path congruent; raw-aligned-but-bias-nonzero; bias=0 alignUp;
     stackAlignment=0 degenerate; AArch64 zero bias; large-shadow
     multi-quanta).
   - test_target_schema.cpp:
     `EntryStackPointerBiasGreaterThanOrEqualAlignmentRejected` —
     validator pin; shipped-cc value pins
     (`sysv_amd64.entryStackPointerBias == 0`,
     `ms_x64.entryStackPointerBias == 8`) extended onto the existing
     ShippedX86_64HasBothSysVAndMsX64 test.
   - test_lk10_entry_slice_c.cpp: `MsX64TrampolinePrologueIsSubRsp0x28`
     pins the 7 leading bytes `48 81 EC 28 00 00 00` of the trampoline
     under PE+ms_x64; `SysvElfTrampolineEmitsNoPrologue` negative pin
     asserts first byte is `0xE8` (CALL) not `0x48` (REX.W SUB) under
     ELF+sysv_amd64. These give bias-regression coverage on EVERY
     host — the existing #if defined(_WIN32) runnable smoke only
     catches it via CreateProcess on Windows.

7. comment-analyzer HIGH + code-architect HIGH (2-agent): two new
   §3.1 rows added to plan 14:
   - `D-LK10-ENTRY-TRAMP-PROLOGUE` ✅ CLOSED 2026-06-02 — full
     bug-mechanism + fix-shape + Win64/SysV/ARM64 derivation history.
   - `D-LK10-ENTRY-ML7-FRAME-BIAS-UNIFY` OPEN — trigger: ML7 cycle
     adds call-site shadow-space allocation for call-making normal
     functions.

**Verified clean**:
- Cross-plan sweep (Explore): zero stale `preCallStackAdjustBytes`
  references in plans/source/configs. Plan 14 §2.13 + plan 11 FF6
  cross-refs consistent. Anchor name uniqueness confirmed.
- Bucket-purity (code-architect Q3): zero `if (target == ...)`
  branches; all per-CPU data via opcodeByMnemonic / registerByName /
  callingConventionByName.
- Algorithm correctness (code-reviewer): alignedSizeWithBias verified
  bit-for-bit across all 5 cases the trampoline + ML7 will exercise.
- Cross-target bias sanity (code-architect Q7): all 3 shipped values
  (ms_x64=8, sysv_amd64=0, aapcs64=0) derived from first-principles
  OS-loader semantics — confirmed correct.

**REJECTED across agents** (all anchored):
- type-design C2 (argument-swap via cc-pull): incompatible with the
  anchored ML7 consumer (ML7 uses post-CALL function-entry bias,
  trampoline uses cc.entryStackPointerBias — distinct biases).
  Anchored D-LK10-ENTRY-TRAMP-PROLOGUE-STRONG-BIAS-TYPE for the
  proper resolution at the 2nd cross-field invariant trigger.
- silent-failure HIGH-2 (cross-format bias-mechanism check): coupling
  mechanism=ByNameImport ↔ shadowSpace>=32 breaks the macOS Mach-O
  case (anchored D-LK10-ENTRY-MACHO-EXIT — also ByNameImport but
  with shadow=0/bias=0). Schema-author responsibility to pair format
  with correct entry cc; existing validator + naming discipline
  catches misconfigurations.
- silent-failure HIGH-4 (stackAlignment=0 + bias!=0 silent return-0):
  anchored as future-ML7-consumer trigger; today unreachable
  (hasAbiInfo predicate forces stackAlignment != 0 via pow2 check).
- silent-failure MEDIUM-5/6 (assemble-empty-without-diagnostic +
  bytes-floor minimum): defense-in-depth anchors.
- type-design C4/C5/C6: optional<bias> + overflow guard + member-form
  — all anchored for triggers.
- code-architect MULTILOADER + VALIDATOR-ZERO-STACK: future-scenario
  anchors.

**Agnosticism re-verified by all 8 agents**: zero `if (target == ...)`,
zero `if (format == ...)`, zero hardcoded byte sequences, zero
language-specific assumptions. Trampoline emitter reads cc primitives,
formula derives the prologue, assembler encodes the `sub` opcode per
the target's encoding row. ARM64 lands via D-LK10-ENTRY-ARM64 with
zero emitter changes.

141/141 ctest (was 141/141; +15 new tests landed in the existing
binaries). CLI e2e re-verified: `int42.c` → `dss-code-prime.exe
--compile --output build/out` → `int42.exe` → Windows runs it →
**exit code 42**.

+++ src/link/entry_trampoline.cpp
@@ -49,6 +49,12 @@ namespace {
     // Walker-side synthesized name convention: `sym_<id>` on
     // ELF/PE; `_sym_<id>` on Mach-O. Match either form (real-name
     // resolution closes with D-LK1-1 / LK7).
+    //
+    // D-LK10-ENTRY-SYNTH-PREFIX-SCHEMA: the two prefix strings are
+    // hardcoded here pending move into the format schema (see plan
+    // 14 §3.1). Closure trigger: 4th format declares processExit OR
+    // D-LK1-1 lands real-symbol-name preservation through the
+    // emit pipeline.
     for (auto const& fn : module.functions) {
         std::string const elfPeName =
             "sym_"  + std::to_string(fn.symbol.v);
@@ -145,9 +151,10 @@ bool injectEntryTrampoline(AssembledModule&          module,
     // Look up all needed opcodes from the target schema (Slice A
     // ships `syscall`, `call_indirect_via_extern`, and the existing
     // `call` / `mov` / `unreachable` opcodes). `sub` is required
-    // ONLY when the format declares a non-zero
-    // `processExit.preCallStackAdjustBytes` (the ABI-prologue arm
-    // closing D-LK10-ENTRY-TRAMP-PROLOGUE).
+    // ONLY when alignedSizeWithBias(cc.shadowSpaceBytes,
+    // cc.stackAlignment, cc.entryStackPointerBias) > 0 — i.e. when
+    // the entry cc declares shadow space OR a non-zero process-
+    // entry RSP bias (closes D-LK10-ENTRY-TRAMP-PROLOGUE).
     auto const callOp     = target.opcodeByMnemonic("call");
     auto const movOp      = target.opcodeByMnemonic("mov");
     auto const unreachOp  = target.opcodeByMnemonic("unreachable");
@@ -257,10 +264,13 @@ bool injectEntryTrampoline(AssembledModule&          module,
     //
     //    No restoration is emitted — the exit mechanism never
     //    returns (the trampoline ends in `unreachable` / `ud2`).
+    // Integral promotion: cc fields are uint16_t; the helper takes
+    // uint32_t. Explicit casts would obscure that this is widening,
+    // not narrowing.
     std::uint32_t const adjustBytes = alignedSizeWithBias(
-        static_cast<std::uint32_t>(cc->shadowSpaceBytes),
-        static_cast<std::uint32_t>(cc->stackAlignment),
-        static_cast<std::uint32_t>(cc->entryStackPointerBias));
+        cc->shadowSpaceBytes,
+        cc->stackAlignment,
+        cc->entryStackPointerBias);
     if (adjustBytes > 0) {
         if (!cc->stackPointer.has_value()) {
             emit(reporter, DiagnosticCode::K_NoMatchingObjectFormat,
